chore(sdk): remove api/v1/types/deployment_operator_v1alpha1
