+++ leptonai/api/v1/types/affinity.py
