--- leptonai/api/v1/types/deployment_operator_v1alpha1/__init__.py
@@ -1,4 +1,0 @@
-"""
-This implements the more detailed types for deployment, jobs etc. You do not
-need to directly use them - the top level leptonai.api.v1.types suffices.
-"""
