--- leptonai/api/v1/types/deployment_operator_v1alpha1/photon.py
@@ -1,12 +1,0 @@
-from pydantic import BaseModel
-from typing import Optional, Dict, List
-
-
-class PhotonDeploymentTemplate(BaseModel):
-    resource_shape: Optional[str] = None
-    env: Optional[Dict[str, str]] = None
-    secret: Optional[List[str]] = None
-
-
-class PhotonStatus(BaseModel):
-    system_photon: Optional[bool] = None
